More fixes
Updates stats for enemies and towers... again
Did caden's tower sprite update
Had to redo game manager into scenetraveler
Made sure end game and win game menus appear
