+++ Assets/CustomPrefabs/Ally/ArcherTowerLevel1.prefab
@@ -48,7 +48,7 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   level: 1
-  range: 10
+  range: 16
   damage: 50
   fireRate: 1
   enemiesInRange: []
@@ -463,8 +463,8 @@ BoxCollider:
   m_IsTrigger: 0
   m_Enabled: 1
   serializedVersion: 2
-  m_Size: {x: 3.1787548, y: 9.159827, z: 3.082975}
-  m_Center: {x: 0, y: 4.5799136, z: -0.0000006556511}
+  m_Size: {x: 3.1787548, y: 6.3, z: 3.082975}
+  m_Center: {x: 0, y: 3.2, z: -0.0000006556511}
 --- !u!114 &692976675430934767
 MonoBehaviour:
   m_ObjectHideFlags: 0
