+++ Assets/CustomPrefabs/TowerDefense/UI/SceneTraveler.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 4848492c0f2b47a49938422b259080d1
+guid: 8d5c17ff169ea5f499fd0f355e0f8309
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
