+++ Assets/CustomPrefabs/TowerDefense/UI/SceneTraveler.prefab
@@ -1,6 +1,6 @@
 %YAML 1.1
 %TAG !u! tag:unity3d.com,2011:
---- !u!1 &4306447093899978558
+--- !u!1 &6492941005514859391
 GameObject:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -8,39 +8,38 @@ GameObject:
   m_PrefabAsset: {fileID: 0}
   serializedVersion: 6
   m_Component:
-  - component: {fileID: 4306447093899978552}
-  - component: {fileID: 4306447093899978553}
+  - component: {fileID: 8654905399186867224}
+  - component: {fileID: 5829871471356308336}
   m_Layer: 0
-  m_Name: GameManager
+  m_Name: SceneTraveler
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
   m_IsActive: 1
---- !u!4 &4306447093899978552
+--- !u!4 &8654905399186867224
 Transform:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 4306447093899978558}
-  m_LocalRotation: {x: -0, y: -0, z: -0, w: 1}
-  m_LocalPosition: {x: 958.778, y: 616.2034, z: -111.74474}
+  m_GameObject: {fileID: 6492941005514859391}
+  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
+  m_LocalPosition: {x: 1.1209059, y: -0.3636284, z: -4.123591}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_Children: []
   m_Father: {fileID: 0}
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!114 &4306447093899978553
+--- !u!114 &5829871471356308336
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 4306447093899978558}
+  m_GameObject: {fileID: 6492941005514859391}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 4848492c0f2b47a49938422b259080d1, type: 3}
+  m_Script: {fileID: 11500000, guid: 8d5c17ff169ea5f499fd0f355e0f8309, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  gameOverUI: {fileID: 0}
