+++ Assets/CustomPrefabs/TowerDefense/UI/SceneTraveler.prefab.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 4f635074f086d8743b084f2ca1e5062d
+guid: d16a52f3b28a0ac498db4ac0b852e300
 PrefabImporter:
   externalObjects: {}
   userData:
